# api/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertIn": {
        "description": "Whether this feed should page someone when it goes quiet.",
        "properties": {
          "armed": {
            "title": "Armed",
            "type": "boolean"
          }
        },
        "required": [
          "armed"
        ],
        "title": "AlertIn",
        "type": "object"
      },
...
      "ExpectationIn": {
        "description": "How often this feed should deliver, or null to stop expecting.\n\nNull rather than a second endpoint: clearing only returns the feed to its\nRedash schedule, and a DELETE beside a PUT would read as destructive.",
        "properties": {
          "expectedIntervalSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedintervalseconds"
          }
        },
        "required": [
          "expectedIntervalSeconds"
        ],
        "title": "ExpectationIn",
        "type": "object"
      },
...
      "FeedOut": {
        "description": "One upstream capture, as the Feed Health board reads it.\n\nA feed here IS a scheduled Redash query that captures into the historical\nwarehouse, so the id is the table it captures into and it matches the\ndataset of the same id one to one. That is structural rather than a\nconvention worth breaking: `get_or_create_table_name` keys the registry on\nquery_id and hands back exactly one table, so `dataset_count` cannot be\nanything but 1 while that holds.",
        "properties": {
          "alertId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alertid"
          },
          "cadence": {
            "title": "Cadence",
            "type": "string"
          },
          "cadenceSource": {
            "enum": [
              "declared",
              "schedule",
              "none"
            ],
            "title": "Cadencesource",
            "type": "string"
          },
          "datasetCount": {
            "title": "Datasetcount",
            "type": "integer"
          },
          "expectedIntervalSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedintervalseconds"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastReceivedAt": {
            "title": "Lastreceivedat",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "fresh",
              "stale"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "source",
          "cadence",
          "cadenceSource",
          "expectedIntervalSeconds",
          "alertId",
          "lastReceivedAt",
          "status",
          "datasetCount"
        ],
        "title": "FeedOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "veodyn-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/feeds": {
      "get": {
        "description": "Every scheduled capture the warehouse holds, with its cadence and source.\n\nThe two Redash lookups run as the CALLER, not as the service account, so a\ndata source or a query the reader cannot see goes unnamed rather than named\nfor them. Both are best-effort by construction: they label a list the\nwarehouse already produced, and neither can empty it.",
        "operationId": "list_feeds_feeds_get",
        "parameters": [
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedOut"
                  },
                  "title": "Response List Feeds Feeds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feeds",
        "tags": [
          "feeds"
        ]
      }
    },
    "/feeds/{feed_id}/alert": {
      "put": {
        "description": "Arm or disarm the late alert on a feed.\n\nRequires a declared expectation, because the alert's threshold is two of\nthose periods. Unlike the expectation, this IS checked against the catalog:\narming writes a query against a real table on a real data source.",
        "operationId": "put_alert_feeds__feed_id__alert_put",
        "parameters": [
          {
            "in": "path",
            "name": "feed_id",
            "required": true,
            "schema": {
              "title": "Feed Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Alert",
        "tags": [
          "feeds"
        ]
      }
    },
    "/feeds/{feed_id}/expectation": {
      "put": {
        "description": "Declare how often a capture is expected to deliver.\n\nNOT checked against the catalog first: a feed id is a warehouse table name,\nand an expectation for a table that has not appeared yet, or has briefly\ndropped out of the registry, is a harmless row that starts working the moment\nit is back. Any authenticated member of the org may set one, since it changes\nno data and no permission, only when this org's board calls a feed late.",
        "operationId": "put_expectation_feeds__feed_id__expectation_put",
        "parameters": [
          {
            "in": "path",
            "name": "feed_id",
            "required": true,
            "schema": {
              "title": "Feed Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Expectation",
        "tags": [
          "feeds"
        ]
      }
    },
...
    "/public/feeds/{slug}": {
      "get": {
        "description": "The current artifact for one public feed, as raw GTFS-Realtime bytes.",
        "operationId": "get_public_feed_public_feeds__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/x-protobuf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The feed's current GTFS-Realtime message."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Feed",
        "tags": [
          "public-feeds"
        ]
      }
    },
...
  }
}